revert google doc import
