+++ package.json
@@ -209,7 +209,6 @@
     "express-session": "^1.17.1",
     "feedparser-promised": "^1.2.2",
     "fp-ts": "^2.13.1",
-    "googleapis": "^133.0.0",
     "gpt-3-encoder": "^1.1.3",
     "graphql": "^14.5.8",
     "graphql-date": "^1.0.2",
