+++ packages/lesswrong/components/common/ForumIcon.tsx
@@ -66,7 +66,6 @@ import ArrowCircleIcon from "@heroicons/react/20/solid/ArrowPathRoundedSquareIco
 import FunnelIcon from "@heroicons/react/24/outline/FunnelIcon";
 import BarsArrowDown from "@heroicons/react/24/outline/BarsArrowDownIcon";
 import InformationCircleIcon from '@heroicons/react/24/solid/InformationCircleIcon';
-import ArrowDownOnSquareIcon from '@heroicons/react/24/outline/ArrowDownOnSquareIcon';
 import MuiVolumeUpIcon from "@material-ui/icons/VolumeUp";
 import MuiBookmarkIcon from "@material-ui/icons/Bookmark";
 import MuiBookmarkBorderIcon from "@material-ui/icons/BookmarkBorder";
@@ -216,8 +215,7 @@ export type ForumIconName =
   "InfoCircle" |
   "BarsArrowDown" |
   "LightbulbChat" |
-  "VoteBallot" |
-  "Import"
+  "VoteBallot"
   ;
 
 const ICONS: ForumOptions<Record<ForumIconName, IconComponent>> = {
@@ -307,7 +305,6 @@ const ICONS: ForumOptions<Record<ForumIconName, IconComponent>> = {
     BarsArrowDown: BarsArrowDown,
     LightbulbChat: LightbulbChatIcon,
     VoteBallot: MuiVoteIcon,
-    Import: ArrowDownOnSquareIcon
   },
   default: {
     VolumeUp: SpeakerWaveIcon,
@@ -395,7 +392,6 @@ const ICONS: ForumOptions<Record<ForumIconName, IconComponent>> = {
     BarsArrowDown: BarsArrowDown,
     LightbulbChat: LightbulbChatIcon,
     VoteBallot: MuiVoteIcon,
-    Import: ArrowDownOnSquareIcon
   },
 };
