+++ packages/lesswrong/components/editor/PostSharingSettings.tsx
@@ -13,10 +13,10 @@ import PersonAddIcon from '@material-ui/icons/PersonAdd';
 import { moderationEmail } from '../../lib/publicSettings';
 import { getPostCollaborateUrl } from '../../lib/collections/posts/helpers';
 import { ckEditorName } from './Editor';
+import { isEAForum } from '../../lib/instanceSettings';
 import { isFriendlyUI } from '../../themes/forumTheme';
-import classNames from 'classnames';
 
-const styles = (theme: ThemeType) => ({
+const styles = (theme: ThemeType): JssStyles => ({
   linkSharingPreview: {
     fontFamily: theme.typography.fontFamily,
   },
@@ -42,23 +42,6 @@ const styles = (theme: ThemeType) => ({
   },
   buttonIcon: {
     cursor: "pointer"
-  },
-  friendlyButton: {
-      color: theme.palette.grey[900],
-      backgroundColor: "transparent",
-      padding: "2px 12px",
-      '&:hover': {
-        backgroundColor: theme.palette.grey[200],
-      },
-      '&:disabled': {
-        color: theme.palette.grey[900],
-        backgroundColor: "transparent",
-        opacity: 0.5
-      },
-  },
-  buttonInternalIcon: {
-    width: "20px",
-    marginRight: 6
   },
   disabledIcon: {
     opacity: .35,
@@ -105,7 +88,7 @@ const PostSharingSettings = ({document, formType, value, classes}: {
   value: SharingSettings,
   path: string,
   label: string,
-  classes: ClassesType<typeof styles>
+  classes: ClassesType
 }, context: any) => {
   const {updateCurrentValues, submitForm} = context;
   const {openDialog, closeDialog} = useDialog();
@@ -170,27 +153,17 @@ const PostSharingSettings = ({document, formType, value, classes}: {
     });
   }, [openDialog, closeDialog, formType, document, updateCurrentValues, initialSharingSettings, flash, submitForm]);
 
-  const {LWTooltip, EAButton} = Components;
-
-  const canUseSharing = userCanUseSharing(currentUser)
-
-  return isFriendlyUI ? (
-    <LWTooltip title={canUseSharing ? undefined : noSharePermissionTooltip}>
-      <EAButton
-        className={classes.friendlyButton}
-        onClick={userCanUseSharing(currentUser) ? onClickShare : undefined}
-        disabled={!canUseSharing}
-      >
-        <PostSharingIcon className={classes.buttonInternalIcon} />
-        Share {document.draft ? " this draft" : ""}
-      </EAButton>
-    </LWTooltip>
-  ) : (
-    <LWTooltip title={canUseSharing ? shareTooltip : noSharePermissionTooltip}>
-      <PostSharingIcon
-        className={classNames(classes.buttonIcon, { [classes.disabledIcon]: !canUseSharing })}
-        onClick={userCanUseSharing(currentUser) ? onClickShare : undefined}
-      />
+  const {LWTooltip} = Components;
+  if (!userCanUseSharing(currentUser)) {
+    return (
+      <LWTooltip title={noSharePermissionTooltip}>
+        <PostSharingIcon className={classes.disabledIcon}/>
+      </LWTooltip>
+    )
+  };
+  return (
+    <LWTooltip title={shareTooltip}>
+      <PostSharingIcon className={classes.buttonIcon} onClick={onClickShare}/>
     </LWTooltip>
   );
 }
