+++ packages/lesswrong/components/posts/PostsPage/PostsRevisionMessage.tsx
@@ -25,7 +25,7 @@ const PostsRevisionMessage = ({post, classes}: {
     <div className={classes.root}>
       You are viewing a version of this post published on the <FormatDate date={post.contents.editedAt} format="Do MMM YYYY"/>.
       {" "}
-      <QueryLink query={{revision: undefined}}>This link</QueryLink> will always display the most recent version of the post.
+      <QueryLink query={{revision: undefined}}>This link</QueryLink> will always display the most recent version of the post..
     </div>
   );
 }
