+++ packages/lesswrong/lib/collections/posts/fragments.ts
@@ -523,6 +523,7 @@ registerFragment(`
       ...RevisionDisplay
     }
     myEditorAccess
+    linkSharingKey
   }
 `)
 
@@ -530,6 +531,7 @@ registerFragment(`
   fragment PostsEdit on Post {
     ...PostsDetails
     myEditorAccess
+    linkSharingKey
     version
     coauthorStatuses
     readTimeMinutesOverride
