+++ packages/lesswrong/lib/crud/withMulti.ts
@@ -164,8 +164,11 @@ export function useMulti<
       terms: { ...terms, limit: defaultLimit },
       enableCache, enableTotal, createIfMissing
     },
-    ...extraVariablesValues
-  }), [terms, defaultLimit, enableCache, enableTotal, createIfMissing, extraVariablesValues]);
+    ...(_.pick(extraVariablesValues, Object.keys(extraVariables || {})))
+  }), [
+    terms, defaultLimit, enableCache, enableTotal, createIfMissing,
+    extraVariables, extraVariablesValues,
+  ]);
 
   let effectiveLimit = limit;
   if (!_.isEqual(terms, lastTerms)) {
