+++ packages/lesswrong/lib/publicSettings.ts
@@ -162,5 +162,3 @@ export const auth0ClientSettings = new DatabasePublicSetting<Auth0ClientSettings
 export const requestFeedbackKarmaLevelSetting = new DatabasePublicSetting<number | null>('post.requestFeedbackKarmaLevel', 100);
 
 export const alwaysShowAnonymousReactsSetting = new DatabasePublicSetting<boolean>('voting.eaEmoji.alwaysShowAnonymousReacts', true);
-
-export const hasGoogleDocImportSetting = new DatabasePublicSetting<boolean>('googleDocImport.enabled', false);
