+++ packages/lesswrong/lib/routes.ts
@@ -1478,12 +1478,6 @@ addRoute(
     title: `${taggingNameCapitalSetting.get()} merging tool`
   },
   {
-    name: 'googleServiceAccount',
-    path: '/admin/googleServiceAccount',
-    componentName: 'AdminGoogleServiceAccount',
-    title: `Google Doc import service account`
-  },
-  {
     name: 'recentlyActiveUsers',
     path: '/admin/recentlyActiveUsers',
     componentName: 'RecentlyActiveUsers',
