+++ packages/lesswrong/lib/vulcan-lib/utils.ts
@@ -315,32 +315,12 @@ const allowedTableStyles = {
 };
 
 const allowedMathMLGlobalAttributes = ['mathvariant', 'dir', 'displaystyle', 'scriptlevel'];
-const footnoteAttributes = [
-  'data-footnote-content',
-  'data-footnote-id',
-  'data-footnote-index',
-  'data-footnote-item',
-  'data-footnote-reference',
-  'data-footnote-section',
-  'data-footnote-back-link',
-  'data-footnote-back-link-href',
-]
-const footnoteClasses = [
-  "footnote-content",
-  "footnote-item",
-  "footnote-reference",
-  "footnote-section",
-  "footnote-back-link",
-  "footnotes",
-  "hidden",
-]
 
 export const sanitize = function(s: string): string {
   return sanitizeHtml(s, {
     allowedTags: sanitizeAllowedTags,
     allowedAttributes:  {
       ...sanitizeHtml.defaults.allowedAttributes,
-      '*': footnoteAttributes,
       audio: [ 'controls', 'src', 'style' ],
       img: [ 'src' , 'srcset', 'alt', 'style'],
       figure: ['style', 'class'],
@@ -401,7 +381,6 @@ export const sanitize = function(s: string): string {
       'calendly.com'
     ],
     allowedClasses: {
-      '*': footnoteClasses,
       span: [ 'footnote-reference', 'footnote-label', 'footnote-back-link' ],
       div: [
         'spoilers',
@@ -422,7 +401,7 @@ export const sanitize = function(s: string): string {
         'ck-cta-button-centered'
       ],
       iframe: [ 'thoughtSaverFrame' ],
-      ol: [ 'footnotes', 'footnote-section' ],
+      ol: [ 'footnotes' ],
       li: [ 'footnote-item' ],
     },
     allowedStyles: {
