+++ packages/lesswrong/server/ckEditor/ckEditorApi.ts
@@ -9,13 +9,12 @@ import Users from '../../lib/collections/users/collection';
 import { userGetDisplayName } from '../../lib/collections/users/helpers';
 import { filterNonnull } from '../../lib/utils/typeGuardUtils';
 import { ckEditorBundleVersion } from '../../lib/wrapCkEditor';
-import { buildRevision } from '../editor/make_editable_callbacks';
+import { buildRevision, getLatestRev, getNextVersion, getPrecedingRev, htmlToChangeMetrics } from '../editor/make_editable_callbacks';
 import { createAdminContext, createMutator, Globals, updateMutator } from '../vulcan-lib';
 import { CkEditorUser, CreateDocumentPayload, DocumentResponse, DocumentResponseSchema, UserSchema } from './ckEditorApiValidators';
 import { getCkEditorApiPrefix, getCkEditorApiSecretKey } from './ckEditorServerConfig';
 import { getPostEditorConfig } from './postEditorConfig';
 import CkEditorUserSessions from '../../lib/collections/ckEditorUserSessions/collection';
-import { getLatestRev, getNextVersion, getPrecedingRev, htmlToChangeMetrics } from '../editor/utils';
 
 // TODO: actually implement these in Zod
 interface CkEditorComment {
