+++ packages/lesswrong/server/editor/utils.ts
@@ -1,11 +1,7 @@
+import pick from 'lodash/pick'
 import mjAPI from 'mathjax-node'
+import Revisions from '../../lib/collections/revisions/collection'
 import {isAnyTest, isMigrations} from '../../lib/executionEnvironment'
-import Revisions, { ChangeMetrics } from '../../lib/collections/revisions/collection';
-import { diff } from "../vendor/node-htmldiff/htmldiff";
-import { cheerioParse } from '../utils/htmlUtil'
-import cheerio from 'cheerio'
-import pick from "lodash/pick";
-import { extractVersionsFromSemver } from '../../lib/editor/utils';
 
 export const trimLatexAndAddCSS = (dom: any, css: string) => {
   // Remove empty paragraphs
@@ -102,61 +98,6 @@ export const preProcessLatex = async (content: AnyBecauseTodo) => {
 
   return content;
 }
-
-const countCharsInTag = (parsedHtml: cheerio.Root, tagName: string): number => {
-  const instancesOfTag = parsedHtml(tagName);
-  let cumulative = 0;
-  for (let i=0; i<instancesOfTag.length; i++) {
-    const tag = instancesOfTag[i];
-    const text = cheerio(tag).text();
-    cumulative += text.length;
-  }
-  return cumulative;
-}
-
-export const htmlToChangeMetrics = (oldHtml: string, newHtml: string): ChangeMetrics => {
-  const htmlDiff = diff(oldHtml, newHtml);
-  const parsedHtml = cheerioParse(htmlDiff);
-
-  /// Given an HTML diff, where added sections are marked with <ins> and <del>
-  /// tags, count the number of chars added and removed. This is used for providing
-  /// a quick distinguisher between small and large changes, on revision history
-  /// lists.
-  const insertedChars = countCharsInTag(parsedHtml, "ins");
-  const removedChars = countCharsInTag(parsedHtml, "del");
-
-  return { added: insertedChars, removed: removedChars };
-}
-
-export async function getLatestRev(documentId: string, fieldName: string): Promise<DbRevision|null> {
-  return await Revisions.findOne({documentId: documentId, fieldName}, {sort: {editedAt: -1}})
-}
-
-/// Given a revision, return the last revision of the same document/field prior
-/// to it (null if the revision is the first).
-export async function getPrecedingRev(rev: DbRevision): Promise<DbRevision|null> {
-  return await Revisions.findOne(
-    {documentId: rev.documentId, fieldName: rev.fieldName, editedAt: {$lt: rev.editedAt}},
-    {sort: {editedAt: -1}}
-  );
-}
-
-export function getNextVersion(previousRevision: DbRevision | null, updateType: DbRevision['updateType'] = 'minor', isDraft: boolean) {
-  const { major, minor, patch } = extractVersionsFromSemver(previousRevision?.version || "1.0.0")
-  switch (updateType) {
-    case "patch":
-      return `${major}.${minor}.${patch + 1}`
-    case "minor":
-      return `${major}.${minor + 1}.0`
-    case "major":
-      return `${major+1}.0.0`
-    case "initial":
-      return isDraft ? '0.1.0' : '1.0.0'
-    default:
-      throw new Error("Invalid updateType, must be one of 'patch', 'minor' or 'major'")
-  }
-}
-
 
 const revisionFieldsToCopy: (keyof DbRevision)[] = [
   "originalContents",
